Add files via upload
Powershell script to send an email out upon an event. Run the commented code part first to save the creds as an exported xml file. Then set up a Task Scheduler to go off and trigger this script via powershell.exe when an event occurs, i.e. a file has been blocked by FSRM or a user is added to Domain Admins group